moved serial and usb speed getter from usb_boot.c to platformDeviceControl.c
