+++ shared/include/XLinkPlatform.h
@@ -55,6 +55,9 @@ int XLinkPlatformBootMemoryRemote(deviceDesc_t* deviceDesc, uint8_t* buffer, lon
 int XLinkPlatformBootRemote(deviceDesc_t* deviceDesc, const char* binaryPath);
 int XLinkPlatformConnect(const char* devPathRead, const char* devPathWrite,
                          XLinkProtocol_t protocol, void** fd);
+
+UsbSpeed_t get_usb_speed();
+char* get_mx_serial();
 #endif // __PC__
 
 int XLinkPlatformCloseRemote(xLinkDeviceHandle_t* deviceHandle);
