+++ shared/src/XLinkDevice.c
@@ -20,7 +20,6 @@
 #endif
 #include "XLinkLog.h"
 #include "XLinkStringUtils.h"
-#include "usb_boot.h"
 
 #define MAX_PATH_LENGTH (255)
 
@@ -214,17 +213,12 @@ XLinkError_t XLinkConnect(XLinkHandler_t* handler)
     link->peerState = XLINK_UP;
     #if (!defined(_WIN32) && !defined(_WIN64) )
         link->usbConnSpeed = get_usb_speed();
-        printf( "Usb Speed:%d\n", link->usbConnSpeed);
-    #else
-        link->usbConnSpeed = X_LINK_USB_UNKNOWN;
-    #endif
-
-    #if (!defined(_WIN32) && !defined(_WIN64) )
         mv_strcpy(link->mxSerialId, XLINK_MAX_MXID, get_mx_serial());
     #else
-        link->mxSerialId = "UNKNOWN";
+        link->usbConnSpeed = XLINK_USB_SPEED_UNKNOWN;
+        mv_strcpy(link->mxSerialId, XLINK_MAX_MXID, "UNKNOWN");
     #endif
-    
+
     link->hostClosedFD = 0;
     handler->linkId = link->id;
     return X_LINK_SUCCESS;
